Work as of 1/11 2:30
Added in Circles, changed handling of circles to jquery, need to add
more circles and see if they actually make a difference with
calibration, I think they might not be helping because they are placed
on top of the plane. Got the mask up, now I need to get the video feed
up.